Include default PDB in Python package
Unconditional designs fail when `inference.input_pdb` is not specified
and the script is ran from outside repository.

This is because `model_runners.py` had a hard-coded relative path
to one of the example PDBs. Unconditional design ignores that PDB
through the contig spec, but the script still fails if it cannot resolve
the file path.

This commit adds the PDB inside an `rfdiffusion.data` namespace package.

Better solution would be to re-work the code, but this is simpler.